+++ inbas-reporting-tool/src/main/webapp/js/components/mixin/MessageMixin.js
@@ -1,6 +1,3 @@
-/**
- * @jsx
- */
 'use strict';
 
 var React = require('react');
@@ -32,6 +29,7 @@ var MessageMixin = {
                 text: text
             }
         });
+        setTimeout(() => this.dismissMessage(), this.dismissInterval);
     },
 
     showSuccessMessage: function (text) {
@@ -45,14 +43,11 @@ var MessageMixin = {
     },
 
     renderMessage: function () {
-        return this.state.message ? (
-            <div className='form-group'>
-                <Alert bsStyle={this.state.message.type} onDismiss={this.dismissMessage}
-                       dismissAfter={this.dismissInterval}>
-                    <p>{this.state.message.text}</p>
-                </Alert>
-            </div>
-        ) : null;
+        return this.state.message ? <div className='form-group'>
+            <Alert bsStyle={this.state.message.type} onDismiss={this.dismissMessage}>
+                <p>{this.state.message.text}</p>
+            </Alert>
+        </div> : null;
     }
 };