+++ inbas-reporting-tool/ReleaseNotes.md
@@ -1,5 +1,9 @@
 # INBAS Reporting Tool Release Notes
 
+### 0.7.8 - October 4, 2016
+- Upgraded to react-bootstrap 0.30.3.
+- Upgraded all the related components - datetime picker, typeahead, semforms.
+
 ### 0.7.4 - September 29, 2016
 - Report type filter buttons above the reports table (Feature #173).
 - The filter buttons do not show when there is only one kind of reports present.